date time

+++ db/migrate/20170617063324_add_amount_to_expenses.rb
@@ -1,5 +1,5 @@
 class AddAmountToExpenses < ActiveRecord::Migration[5.0]
   def change
-    add_column :expenses, :amount, :integer
+    add_column :expenses, :amount, :float
   end
 end
